Separate settings into multiple files, refactored Lyrics management into a dedicated LyricsService, added utility functions to determine real notch height and menu bar height.
