+++ boringNotch/XPCHelperClient/XPCHelperClient.swift
@@ -257,6 +257,3 @@ final class XPCHelperClient: NSObject {
     }
 }
 
-extension Notification.Name {
-    static let accessibilityAuthorizationChanged = Notification.Name("accessibilityAuthorizationChanged")
-}
