+++ boringNotch/components/Settings/Views/AboutView.swift
@@ -1,0 +1,83 @@
+//
+//  AboutView.swift
+//  boringNotch
+//
+//  Created by Richard Kunkli on 07/08/2024.
+//
+
+import Defaults
+import Sparkle
+import SwiftUI
+
+struct About: View {
+    @State private var showBuildNumber: Bool = false
+    let updaterController: SPUStandardUpdaterController
+    @Environment(\.openWindow) var openWindow
+    var body: some View {
+        VStack {
+            Form {
+                Section {
+                    HStack {
+                        Text("Release name")
+                        Spacer()
+                        Text(Defaults[.releaseName])
+                            .foregroundStyle(.secondary)
+                    }
+                    HStack {
+                        Text("Version")
+                        Spacer()
+                        if showBuildNumber {
+                            Text("(\(Bundle.main.buildVersionNumber ?? ""))")
+                                .foregroundStyle(.secondary)
+                        }
+                        Text(Bundle.main.releaseVersionNumber ?? "unkown")
+                            .foregroundStyle(.secondary)
+                    }
+                    .onTapGesture {
+                        withAnimation {
+                            showBuildNumber.toggle()
+                        }
+                    }
+                } header: {
+                    Text("Version info")
+                }
+
+                UpdaterSettingsView(updater: updaterController.updater)
+
+                HStack(spacing: 30) {
+                    Spacer(minLength: 0)
+                    Button {
+                        if let url = URL(string: "https://github.com/TheBoredTeam/boring.notch") {
+                            NSWorkspace.shared.open(url)
+                        }
+                    } label: {
+                        VStack(spacing: 5) {
+                            Image("Github")
+                                .resizable()
+                                .aspectRatio(contentMode: .fit)
+                                .frame(width: 18)
+                            Text("GitHub")
+                        }
+                        .contentShape(Rectangle())
+                    }
+                    Spacer(minLength: 0)
+                }
+                .buttonStyle(PlainButtonStyle())
+            }
+            VStack(spacing: 0) {
+                Divider()
+                Text("Made with 🫶🏻 by not so boring not.people")
+                    .foregroundStyle(.secondary)
+                    .padding(.top, 5)
+                    .padding(.bottom, 7)
+                    .multilineTextAlignment(.center)
+                    .padding(.horizontal, 10)
+            }
+            .frame(maxWidth: .infinity, alignment: .center)
+        }
+        .toolbar {
+            CheckForUpdatesView(updater: updaterController.updater)
+        }
+        .navigationTitle("About")
+    }
+}
