+++ boringNotch/components/Settings/Views/AdvancedSettingsView.swift
@@ -1,0 +1,325 @@
+//
+//  AdvancedSettingsView.swift
+//  boringNotch
+//
+//  Created by Richard Kunkli on 07/08/2024.
+//
+
+import Defaults
+import SwiftUI
+
+struct Advanced: View {
+    @Default(.useCustomAccentColor) var useCustomAccentColor
+    @Default(.customAccentColorData) var customAccentColorData
+    @Default(.extendHoverArea) var extendHoverArea
+    @Default(.showOnLockScreen) var showOnLockScreen
+    @Default(.hideFromScreenRecording) var hideFromScreenRecording
+    
+    @State private var customAccentColor: Color = .accentColor
+    @State private var selectedPresetColor: PresetAccentColor? = nil
+    let icons: [String] = ["logo2"]
+    @State private var selectedIcon: String = "logo2"
+    
+    // macOS accent colors
+    enum PresetAccentColor: String, CaseIterable, Identifiable {
+        case blue = "Blue"
+        case purple = "Purple"
+        case pink = "Pink"
+        case red = "Red"
+        case orange = "Orange"
+        case yellow = "Yellow"
+        case green = "Green"
+        case graphite = "Graphite"
+        
+        var id: String { self.rawValue }
+        
+        var color: Color {
+            switch self {
+            case .blue: return Color(red: 0.0, green: 0.478, blue: 1.0)
+            case .purple: return Color(red: 0.686, green: 0.322, blue: 0.871)
+            case .pink: return Color(red: 1.0, green: 0.176, blue: 0.333)
+            case .red: return Color(red: 1.0, green: 0.271, blue: 0.227)
+            case .orange: return Color(red: 1.0, green: 0.584, blue: 0.0)
+            case .yellow: return Color(red: 1.0, green: 0.8, blue: 0.0)
+            case .green: return Color(red: 0.4, green: 0.824, blue: 0.176)
+            case .graphite: return Color(red: 0.557, green: 0.557, blue: 0.576)
+            }
+        }
+    }
+    
+    var body: some View {
+        Form {
+            Section {
+                VStack(alignment: .leading, spacing: 16) {
+                    // Toggle between system and custom
+                    Picker("Accent color", selection: $useCustomAccentColor) {
+                        Text("System").tag(false)
+                        Text("Custom").tag(true)
+                    }
+                    .pickerStyle(.segmented)
+                    
+                    if !useCustomAccentColor {
+                        // System accent info
+                        VStack(alignment: .leading, spacing: 8) {
+                            HStack(spacing: 12) {
+                                AccentCircleButton(
+                                    isSelected: true,
+                                    color: .accentColor,
+                                    isSystemDefault: true
+                                ) {}
+                                
+                                VStack(alignment: .leading, spacing: 2) {
+                                    Text("Using System Accent")
+                                        .font(.body)
+                                    Text("Your macOS system accent color")
+                                        .font(.caption)
+                                        .foregroundStyle(.secondary)
+                                }
+                                Spacer()
+                            }
+                        }
+                    } else {
+                        // Custom color options
+                        VStack(alignment: .leading, spacing: 12) {
+                            Text("Color Presets")
+                                .font(.caption)
+                                .fontWeight(.semibold)
+                                .foregroundStyle(.secondary)
+                            
+                            HStack(spacing: 12) {
+                                ForEach(PresetAccentColor.allCases) { preset in
+                                    AccentCircleButton(
+                                        isSelected: selectedPresetColor == preset,
+                                        color: preset.color,
+                                        isMulticolor: false
+                                    ) {
+                                        selectedPresetColor = preset
+                                        customAccentColor = preset.color
+                                        saveCustomColor(preset.color)
+                                        forceUiUpdate()
+                                    }
+                                }
+                                Spacer()
+                            }
+                            
+                            Divider()
+                                .padding(.vertical, 4)
+                            
+                            // Custom color picker
+                            HStack(spacing: 12) {
+                                VStack(alignment: .leading, spacing: 2) {
+                                    Text("Pick a Color")
+                                        .font(.body)
+                                    Text("Choose any color")
+                                        .font(.caption)
+                                        .foregroundStyle(.secondary)
+                                }
+                                
+                                Spacer()
+                                
+                                ColorPicker(selection: Binding(
+                                    get: { customAccentColor },
+                                    set: { newColor in
+                                        customAccentColor = newColor
+                                        selectedPresetColor = nil
+                                        saveCustomColor(newColor)
+                                        forceUiUpdate()
+                                    }
+                                ), supportsOpacity: false) {
+                                    ZStack {
+                                        Circle()
+                                            .fill(customAccentColor)
+                                            .frame(width: 32, height: 32)
+                                        
+                                        if selectedPresetColor == nil {
+                                            Circle()
+                                                .strokeBorder(.primary.opacity(0.3), lineWidth: 2)
+                                                .frame(width: 32, height: 32)
+                                        }
+                                    }
+                                }
+                                .labelsHidden()
+                            }
+                        }
+                    }
+                }
+                .padding(.vertical, 4)
+            } header: {
+                Text("Accent color")
+            } footer: {
+                Text("Choose between your system accent color or customize it with your own selection.")
+                    .multilineTextAlignment(.trailing)
+                    .foregroundStyle(.secondary)
+                    .font(.caption)
+            }
+            .onAppear {
+                initializeAccentColorState()
+            }
+            
+            Section {
+                Defaults.Toggle(key: .enableShadow) {
+                    Text("Enable window shadow")
+                }
+                Defaults.Toggle(key: .cornerRadiusScaling) {
+                    Text("Scale corner radius for closed notch")
+                }
+            } header: {
+                Text("Window Appearance")
+            }
+            
+            Section {
+                HStack {
+                    ForEach(icons, id: \.self) { icon in
+                        Spacer()
+                        VStack {
+                            Image(icon)
+                                .resizable()
+                                .frame(width: 80, height: 80)
+                                .background(
+                                    RoundedRectangle(cornerRadius: 20, style: .circular)
+                                        .strokeBorder(
+                                            icon == selectedIcon ? Color.effectiveAccent : .clear,
+                                            lineWidth: 2.5
+                                        )
+                                )
+
+                            Text("Default")
+                                .fontWeight(.medium)
+                                .font(.caption)
+                                .foregroundStyle(icon == selectedIcon ? .white : .secondary)
+                                .padding(.horizontal, 10)
+                                .padding(.vertical, 3)
+                                .background(
+                                    Capsule()
+                                        .fill(icon == selectedIcon ? Color.effectiveAccent : .clear)
+                                )
+                        }
+                        .onTapGesture {
+                            withAnimation {
+                                selectedIcon = icon
+                            }
+                            NSApp.applicationIconImage = NSImage(named: icon)
+                        }
+                        Spacer()
+                    }
+                }
+                .disabled(true)
+            } header: {
+                HStack {
+                    Text("App icon")
+                    customBadge(text: "Coming soon")
+                }
+            }
+            
+            Section {
+                Defaults.Toggle(key: .extendHoverArea) {
+                    Text("Extend hover area")
+                }
+                Defaults.Toggle(key: .hideTitleBar) {
+                    Text("Hide title bar")
+                }
+                Defaults.Toggle(key: .showOnLockScreen) {
+                    Text("Show notch on lock screen")
+                }
+                Defaults.Toggle(key: .hideFromScreenRecording) {
+                    Text("Hide from screen recording")
+                }
+                Defaults.Toggle(key: .hideNonNotchedFromMissionControl) {
+                    Text("Hide windows on non-notch displays from Mission Control")
+                }
+            } header: {
+                Text("Window Behavior")
+            }
+        }
+        .accentColor(.effectiveAccent)
+        .navigationTitle("Advanced")
+        .onAppear {
+            loadCustomColor()
+        }
+    }
+    
+    private func forceUiUpdate() {
+        // Force refresh the UI
+        DispatchQueue.main.async {
+            NotificationCenter.default.post(name: .accentColorChanged, object: nil)
+        }
+    }
+    
+    private func saveCustomColor(_ color: Color) {
+        let nsColor = NSColor(color)
+        if let colorData = try? NSKeyedArchiver.archivedData(withRootObject: nsColor, requiringSecureCoding: false) {
+            Defaults[.customAccentColorData] = colorData
+            forceUiUpdate()
+        }
+    }
+    
+    private func loadCustomColor() {
+        if let colorData = Defaults[.customAccentColorData],
+           let nsColor = try? NSKeyedUnarchiver.unarchivedObject(ofClass: NSColor.self, from: colorData) {
+            customAccentColor = Color(nsColor: nsColor)
+            
+            // Check if loaded color matches a preset
+            selectedPresetColor = nil
+            for preset in PresetAccentColor.allCases {
+                if colorsAreEqual(Color(nsColor: nsColor), preset.color) {
+                    selectedPresetColor = preset
+                    break
+                }
+            }
+        }
+    }
+    
+    private func colorsAreEqual(_ color1: Color, _ color2: Color) -> Bool {
+        let nsColor1 = NSColor(color1).usingColorSpace(.sRGB) ?? NSColor(color1)
+        let nsColor2 = NSColor(color2).usingColorSpace(.sRGB) ?? NSColor(color2)
+        
+        return abs(nsColor1.redComponent - nsColor2.redComponent) < 0.01 &&
+               abs(nsColor1.greenComponent - nsColor2.greenComponent) < 0.01 &&
+               abs(nsColor1.blueComponent - nsColor2.blueComponent) < 0.01
+    }
+    
+    private func initializeAccentColorState() {
+        if !useCustomAccentColor {
+            selectedPresetColor = nil // Multicolor is selected when useCustomAccentColor is false
+        } else {
+            loadCustomColor()
+        }
+    }
+}
+
+// MARK: - Accent Circle Button Component
+struct AccentCircleButton: View {
+    let isSelected: Bool
+    let color: Color
+    var isSystemDefault: Bool = false
+    var isMulticolor: Bool = false
+    let action: () -> Void
+    
+    var body: some View {
+        Button(action: action) {
+            ZStack {
+                // Color circle
+                Circle()
+                    .fill(color)
+                    .frame(width: 32, height: 32)
+                
+                // Subtle border
+                Circle()
+                    .strokeBorder(Color.primary.opacity(0.15), lineWidth: 1)
+                    .frame(width: 32, height: 32)
+                
+                // Apple-style highlight ring around the middle when selected
+                if isSelected {
+                    Circle()
+                        .strokeBorder(
+                            Color.white.opacity(0.5),
+                            lineWidth: 2
+                        )
+                        .frame(width: 28, height: 28)
+                }
+            }
+        }
+        .buttonStyle(.plain)
+        .help(isSystemDefault ? "Use your macOS system accent color" : "")
+    }
+}
