+++ boringNotch/components/Settings/Views/BatterySettingsView.swift
@@ -1,0 +1,43 @@
+//
+//  BatterySettingsView.swift
+//  boringNotch
+//
+//  Created by Richard Kunkli on 07/08/2024.
+//
+
+import Defaults
+import SwiftUI
+
+struct Charge: View {
+    var body: some View {
+        Form {
+            Section {
+                Defaults.Toggle(key: .showBatteryIndicator) {
+                    Text("Show battery indicator")
+                }
+                Defaults.Toggle(key: .showPowerStatusNotifications) {
+                    Text("Show power status notifications")
+                }
+            } header: {
+                Text("General")
+            }
+            Section {
+                Defaults.Toggle(key: .showBatteryPercentage) {
+                    Text("Show battery percentage")
+                }
+                Defaults.Toggle(key: .showPowerStatusIcons) {
+                    Text("Show power status icons")
+                }
+            } header: {
+                Text("Battery Information")
+            }
+        }
+        .onAppear {
+            Task { @MainActor in
+                await XPCHelperClient.shared.isAccessibilityAuthorized()
+            }
+        }
+        .accentColor(.effectiveAccent)
+        .navigationTitle("Battery")
+    }
+}
