+++ boringNotch/components/Settings/Views/GeneralSettingsView.swift
@@ -1,0 +1,237 @@
+//
+//  GeneralSettingsView.swift
+//  boringNotch
+//
+//  Created by Richard Kunkli on 07/08/2024.
+//
+
+import Defaults
+import LaunchAtLogin
+import SwiftUI
+
+struct GeneralSettings: View {
+    @State private var screens: [(uuid: String, name: String)] = NSScreen.screens.compactMap { screen in
+        guard let uuid = screen.displayUUID else { return nil }
+        return (uuid, screen.localizedName)
+    }
+    @EnvironmentObject var vm: BoringViewModel
+    @ObservedObject var coordinator = BoringViewCoordinator.shared
+
+    @Default(.mirrorShape) var mirrorShape
+    @Default(.gestureSensitivity) var gestureSensitivity
+    @Default(.minimumHoverDuration) var minimumHoverDuration
+    @Default(.nonNotchHeight) var nonNotchHeight
+    @Default(.nonNotchHeightMode) var nonNotchHeightMode
+    @Default(.notchHeight) var notchHeight
+    @Default(.notchHeightMode) var notchHeightMode
+    @Default(.showOnAllDisplays) var showOnAllDisplays
+    @Default(.automaticallySwitchDisplay) var automaticallySwitchDisplay
+    @Default(.enableGestures) var enableGestures
+    @Default(.openNotchOnHover) var openNotchOnHover
+
+    var body: some View {
+        Form {
+            Section {
+                Toggle(isOn: Binding(
+                    get: { Defaults[.menubarIcon] },
+                    set: { Defaults[.menubarIcon] = $0 }
+                )) {
+                    Text("Show menu bar icon")
+                }
+                .tint(.effectiveAccent)
+                LaunchAtLogin.Toggle("Launch at login")
+                Defaults.Toggle(key: .showOnAllDisplays) {
+                    Text("Show on all displays")
+                }
+                .onChange(of: showOnAllDisplays) {
+                    NotificationCenter.default.post(
+                        name: Notification.Name.showOnAllDisplaysChanged, object: nil)
+                }
+                Picker("Preferred display", selection: $coordinator.preferredScreenUUID) {
+                    ForEach(screens, id: \.uuid) { screen in
+                        Text(screen.name).tag(screen.uuid as String?)
+                    }
+                }
+                .onChange(of: NSScreen.screens) {
+                    screens = NSScreen.screens.compactMap { screen in
+                        guard let uuid = screen.displayUUID else { return nil }
+                        return (uuid, screen.localizedName)
+                    }
+                }
+                .disabled(showOnAllDisplays)
+                
+                Defaults.Toggle(key: .automaticallySwitchDisplay) {
+                    Text("Automatically switch displays")
+                }
+                    .onChange(of: automaticallySwitchDisplay) {
+                        NotificationCenter.default.post(
+                            name: Notification.Name.automaticallySwitchDisplayChanged, object: nil)
+                    }
+                    .disabled(showOnAllDisplays)
+            } header: {
+                Text("System features")
+            }
+
+            Section {
+                Picker(
+                    selection: $notchHeightMode,
+                    label:
+                        Text("Notch height on notch displays")
+                ) {
+                    Text("Match real notch height")
+                        .tag(WindowHeightMode.matchRealNotchSize)
+                    Text("Match menu bar height")
+                        .tag(WindowHeightMode.matchMenuBar)
+                    Text("Custom height")
+                        .tag(WindowHeightMode.custom)
+                }
+                .onChange(of: notchHeightMode) {
+                    switch notchHeightMode {
+                    case .matchRealNotchSize:
+                        // Get the actual notch height from the built-in display
+                        notchHeight = getRealNotchHeight()
+                    case .matchMenuBar:
+                        notchHeight = 43
+                    case .custom:
+                        notchHeight = 38
+                    }
+                    NotificationCenter.default.post(
+                        name: Notification.Name.notchHeightChanged, object: nil)
+                }
+                if notchHeightMode == .custom {
+                    Slider(value: $notchHeight, in: 15...45, step: 1) {
+                        Text("Custom notch size - \(notchHeight, specifier: "%.0f")")
+                    }
+                    .onChange(of: notchHeight) {
+                        NotificationCenter.default.post(
+                            name: Notification.Name.notchHeightChanged, object: nil)
+                    }
+                }
+                Picker("Notch height on non-notch displays", selection: $nonNotchHeightMode) {
+                    Text("Match menubar height")
+                        .tag(WindowHeightMode.matchMenuBar)
+                    Text("Custom height")
+                        .tag(WindowHeightMode.custom)
+                }
+                .onChange(of: nonNotchHeightMode) {
+                    switch nonNotchHeightMode {
+                    case .matchMenuBar:
+                        nonNotchHeight = 23
+                    case .matchRealNotchSize, .custom:
+                        nonNotchHeight = 23
+                    }
+                    NotificationCenter.default.post(
+                        name: Notification.Name.notchHeightChanged, object: nil)
+                }
+                if nonNotchHeightMode == .custom {
+                    // Custom binding to skip values 1-14 (jump from 0 to 10)
+                    let sliderValue = Binding<Double>(
+                        get: { 
+                            nonNotchHeight == 0 ? 0 : nonNotchHeight - 14
+                        },
+                        set: { newValue in
+                            let oldValue = nonNotchHeight
+                            nonNotchHeight = newValue == 0 ? 0 : newValue + 14
+                            if oldValue != nonNotchHeight {
+                                NotificationCenter.default.post(
+                                    name: Notification.Name.notchHeightChanged, object: nil)
+                            }
+                        }
+                    )
+                    
+                    Slider(value: sliderValue, in: 0...26, step: 1) {
+                        Text("Custom notch size - \(nonNotchHeight, specifier: "%.0f")")
+                    }
+                }
+            } header: {
+                Text("Notch sizing")
+            }
+
+            NotchBehaviour()
+
+            gestureControls()
+        }
+        .toolbar {
+            Button("Quit app") {
+                NSApp.terminate(self)
+            }
+            .controlSize(.extraLarge)
+        }
+        .accentColor(.effectiveAccent)
+        .navigationTitle("General")
+        .onChange(of: openNotchOnHover) {
+            if !openNotchOnHover {
+                enableGestures = true
+            }
+        }
+    }
+
+    @ViewBuilder
+    func gestureControls() -> some View {
+        Section {
+            Defaults.Toggle(key: .enableGestures) {
+                Text("Enable gestures")
+            }
+                .disabled(!openNotchOnHover)
+            if enableGestures {
+                Toggle("Change media with horizontal gestures", isOn: .constant(false))
+                    .disabled(true)
+                Defaults.Toggle(key: .closeGestureEnabled) {
+                    Text("Close gesture")
+                }
+                Slider(value: $gestureSensitivity, in: 100...300, step: 100) {
+                    HStack {
+                        Text("Gesture sensitivity")
+                        Spacer()
+                        Text(
+                            Defaults[.gestureSensitivity] == 100
+                                ? "High" : Defaults[.gestureSensitivity] == 200 ? "Medium" : "Low"
+                        )
+                        .foregroundStyle(.secondary)
+                    }
+                }
+            }
+        } header: {
+            HStack {
+                Text("Gesture control")
+                customBadge(text: "Beta")
+            }
+        } footer: {
+            Text(
+                "Two-finger swipe up on notch to close, two-finger swipe down on notch to open when **Open notch on hover** option is disabled"
+            )
+            .multilineTextAlignment(.trailing)
+            .foregroundStyle(.secondary)
+            .font(.caption)
+        }
+    }
+
+    @ViewBuilder
+    func NotchBehaviour() -> some View {
+        Section {
+            Defaults.Toggle(key: .openNotchOnHover) {
+                Text("Open notch on hover")
+            }
+            Defaults.Toggle(key: .enableHaptics) {
+                    Text("Enable haptic feedback")
+            }
+            Toggle("Remember last tab", isOn: $coordinator.openLastTabByDefault)
+            if openNotchOnHover {
+                Slider(value: $minimumHoverDuration, in: 0...1, step: 0.1) {
+                    HStack {
+                        Text("Hover delay")
+                        Spacer()
+                        Text("\(minimumHoverDuration, specifier: "%.1f")s")
+                            .foregroundStyle(.secondary)
+                    }
+                }
+                .onChange(of: minimumHoverDuration) {
+                    NotificationCenter.default.post(
+                        name: Notification.Name.notchHeightChanged, object: nil)
+                }
+            }
+        } header: {
+            Text("Notch behavior")
+        }
+    }
+}
