+++ boringNotch/components/Settings/Views/HUDSettingsView.swift
@@ -1,0 +1,138 @@
+//
+//  HUDSettingsView.swift
+//  boringNotch
+//
+//  Created by Richard Kunkli on 07/08/2024.
+//
+
+import Defaults
+import SwiftUI
+
+struct HUD: View {
+    @EnvironmentObject var vm: BoringViewModel
+    @Default(.inlineHUD) var inlineHUD
+    @Default(.enableGradient) var enableGradient
+    @Default(.optionKeyAction) var optionKeyAction
+    @Default(.hudReplacement) var hudReplacement
+    @ObservedObject var coordinator = BoringViewCoordinator.shared
+    @State private var accessibilityAuthorized = false
+    
+    var body: some View {
+        Form {
+            Section {
+                HStack {
+                    VStack(alignment: .leading, spacing: 2) {
+                        Text("Replace system HUD")
+                            .font(.headline)
+                        Text("Replaces the standard macOS volume, display brightness, and keyboard brightness HUDs with a custom design.")
+                            .font(.subheadline)
+                            .foregroundStyle(.secondary)
+                            .fixedSize(horizontal: false, vertical: true)
+                    }
+                    Spacer(minLength: 40)
+                    Defaults.Toggle("", key: .hudReplacement)
+                    .labelsHidden()
+                    .toggleStyle(.switch)
+                    .controlSize(.large)
+                    .disabled(!accessibilityAuthorized)
+                }
+                
+                if !accessibilityAuthorized {
+                    VStack(alignment: .leading, spacing: 8) {
+                        Text("Accessibility access is required to replace the system HUD.")
+                            .font(.subheadline)
+                            .foregroundStyle(.secondary)
+
+                        HStack(spacing: 12) {
+                            Button("Request Accessibility") {
+                                XPCHelperClient.shared.requestAccessibilityAuthorization()
+                            }
+                            .buttonStyle(.borderedProminent)
+                        }
+                    }
+                    .padding(.top, 6)
+                }
+            }
+            
+            Section {
+                Picker("Option key behaviour", selection: $optionKeyAction) {
+                    ForEach(OptionKeyAction.allCases) { opt in
+                        Text(opt.rawValue).tag(opt)
+                    }
+                }
+                
+                Picker("Progress bar style", selection: $enableGradient) {
+                    Text("Hierarchical")
+                        .tag(false)
+                    Text("Gradient")
+                        .tag(true)
+                }
+                Defaults.Toggle(key: .systemEventIndicatorShadow) {
+                    Text("Enable glowing effect")
+                }
+                Defaults.Toggle(key: .systemEventIndicatorUseAccent) {
+                    Text("Tint progress bar with accent color")
+                }
+            } header: {
+                Text("General")
+            }
+            .disabled(!hudReplacement)
+            
+            Section {
+                Defaults.Toggle(key: .showOpenNotchHUD) {
+                    Text("Show HUD in open notch")
+                }
+                Defaults.Toggle(key: .showOpenNotchHUDPercentage) {
+                    Text("Show percentage")
+                }
+                .disabled(!Defaults[.showOpenNotchHUD])
+            } header: {
+                HStack {
+                    Text("Open Notch")
+                    customBadge(text: "Beta")
+                }
+            }
+            .disabled(!hudReplacement)
+            
+            Section {
+                Picker("HUD style", selection: $inlineHUD) {
+                    Text("Default")
+                        .tag(false)
+                    Text("Inline")
+                        .tag(true)
+                }
+                .onChange(of: Defaults[.inlineHUD]) {
+                    if Defaults[.inlineHUD] {
+                        withAnimation {
+                            Defaults[.systemEventIndicatorShadow] = false
+                            Defaults[.enableGradient] = false
+                        }
+                    }
+                }
+                
+                Defaults.Toggle(key: .showClosedNotchHUDPercentage) {
+                    Text("Show percentage")
+                }
+            } header: {
+                Text("Closed Notch")
+            }
+            .disabled(!Defaults[.hudReplacement])
+        }
+        .accentColor(.effectiveAccent)
+        .navigationTitle("HUDs")
+        .task {
+            accessibilityAuthorized = await XPCHelperClient.shared.isAccessibilityAuthorized()
+        }
+        .onAppear {
+            XPCHelperClient.shared.startMonitoringAccessibilityAuthorization()
+        }
+        .onDisappear {
+            XPCHelperClient.shared.stopMonitoringAccessibilityAuthorization()
+        }
+        .onReceive(NotificationCenter.default.publisher(for: .accessibilityAuthorizationChanged)) { notification in
+            if let granted = notification.userInfo?["granted"] as? Bool {
+                accessibilityAuthorized = granted
+            }
+        }
+    }
+}
