+++ boringNotch/components/Settings/Views/SettingsHelpers.swift
@@ -1,0 +1,56 @@
+//
+//  SettingsHelpers.swift
+//  boringNotch
+//
+//  Created by Richard Kunkli on 07/08/2024.
+//
+
+import SwiftUI
+
+func proFeatureBadge() -> some View {
+    Text("Upgrade to Pro")
+        .foregroundStyle(Color(red: 0.545, green: 0.196, blue: 0.98))
+        .font(.footnote.bold())
+        .padding(.vertical, 3)
+        .padding(.horizontal, 6)
+        .background(
+            RoundedRectangle(cornerRadius: 4).stroke(
+                Color(red: 0.545, green: 0.196, blue: 0.98), lineWidth: 1))
+}
+
+func comingSoonTag() -> some View {
+    Text("Coming soon")
+        .foregroundStyle(.secondary)
+        .font(.footnote.bold())
+        .padding(.vertical, 3)
+        .padding(.horizontal, 6)
+        .background(Color(nsColor: .secondarySystemFill))
+        .clipShape(.capsule)
+}
+
+func customBadge(text: String) -> some View {
+    Text(text)
+        .foregroundStyle(.secondary)
+        .font(.footnote.bold())
+        .padding(.vertical, 3)
+        .padding(.horizontal, 6)
+        .background(Color(nsColor: .secondarySystemFill))
+        .clipShape(.capsule)
+}
+
+func warningBadge(_ text: String, _ description: String) -> some View {
+    Section {
+        HStack(spacing: 12) {
+            Image(systemName: "exclamationmark.triangle.fill")
+                .font(.system(size: 22))
+                .foregroundStyle(.yellow)
+            VStack(alignment: .leading) {
+                Text(text)
+                    .font(.headline)
+                Text(description)
+                    .foregroundStyle(.secondary)
+            }
+            Spacer()
+        }
+    }
+}
