+++ boringNotch/components/Settings/Views/ShortcutsSettingsView.swift
@@ -1,0 +1,33 @@
+//
+//  ShortcutsSettingsView.swift
+//  boringNotch
+//
+//  Created by Richard Kunkli on 07/08/2024.
+//
+
+import KeyboardShortcuts
+import SwiftUI
+
+struct Shortcuts: View {
+    var body: some View {
+        Form {
+            Section {
+                KeyboardShortcuts.Recorder("Toggle Sneak Peek:", name: .toggleSneakPeek)
+            } header: {
+                Text("Media")
+            } footer: {
+                Text(
+                    "Sneak Peek shows the media title and artist under the notch for a few seconds."
+                )
+                .multilineTextAlignment(.trailing)
+                .foregroundStyle(.secondary)
+                .font(.caption)
+            }
+            Section {
+                KeyboardShortcuts.Recorder("Toggle Notch Open:", name: .toggleNotchOpen)
+            }
+        }
+        .accentColor(.effectiveAccent)
+        .navigationTitle("Shortcuts")
+    }
+}
